feat: add `requiredDeep()`
fix: use `deepmerge` instead of `lodash.merge` internally.

`lodash.merge` is big compare to `deepmerge`.

Change the behavior when the first argument is undefined or null.
This could be breaking but I consider it as not because the type prevent this from happening.
And in normal usecase this should not happen.

The new behavior is more consistent.